labs/Seminar-5.1-STUDENT/requirements.txt
# ──────────────────────────────────────────────────────────────────
# Seminar 5.1 — Python dependencies
# Python 3.10+ required
# Install with: pip install -r requirements.txt
# ──────────────────────────────────────────────────────────────────

# TP2 — HTTP/2 server (ASGI framework + HTTP/2 support)
quart>=0.19.0
hypercorn[h2]>=0.16.0

# TP2 — HTTP/2 client (h2 backend already installed via hypercorn[h2])
httpx>=0.27.0

# TP3 — HTTP/3 / QUIC server and client
aioquic>=1.0.0

# TP4 — gRPC server and client
grpcio>=1.62.0

# TP4 — gRPC stub generation from .proto files (used by setup.sh)
grpcio-tools>=1.62.0

# Shared — TLS certificate generation
cryptography>=41.0.0
